added new query for all projects connected to an org
