+++ schema.graphql
@@ -12,6 +12,7 @@ type Query {
   ikcReport(id: ID!): IKCReport
   researchProject(id: ID!): ResearchProject
   getAllOrgsForProject(projectId: ID!): [PartnerOrg]
+  getAllProjForOrgs(orgId: ID!): [ResearchProject]
 }
 
 type Mutation {
