+++ src/utils/reducers.ts
@@ -1,4 +1,5 @@
 import { Contributor as PrismaContributor, User } from "@prisma/client";
+import prisma from "../../libs/prisma";
 import {
   Contribution,
   Contributor as GraphQLContributor,
@@ -141,6 +142,55 @@ export const getIKCReportArray = (ikcReports): IkcReport[] => {
       },
     ];
   }, []);
+};
+
+export const getResearchProject = async (id) => {
+  const researchProject = await prisma.researchProject.findUnique({
+    where: { id },
+    include: {
+      admins: true,
+      projectPartners: {
+        select: { id: true, name: true, contributors: true },
+      },
+      ikcReports: {
+        select: {
+          id: true,
+          partnerOrgId: true,
+          reportStartDate: true,
+          reportEndDate: true,
+          Contributions: {
+            select: {
+              id: true,
+              contributorId: true,
+              date: true,
+              details: true,
+              hourContribution: true,
+              otherContribution: true,
+            },
+          },
+          submitterId: true,
+          isApproved: true,
+          approverId: true,
+          approvalDate: true,
+        },
+      },
+    },
+  });
+
+  const projectAdmins = getAdminIds(researchProject.admins);
+
+  const projectPartners = getProjectPartners(researchProject.projectPartners);
+  const projectReports = getIKCReportArray(researchProject.ikcReports);
+
+  return {
+    id: researchProject.id,
+    projectTitle: researchProject.projectTitle,
+    startDate: researchProject.startDate,
+    endDate: researchProject.endDate,
+    admins: projectAdmins,
+    projectPartners: projectPartners,
+    ikcReports: projectReports,
+  };
 };
 
 // .reduce((previousValue, currentValue) => {
